Moved `example-x` to `demo-project/x`
