+++ demo-project/build.gradle.kts
@@ -1,0 +1,4 @@
+allprojects {
+    group = "com.github.gmazzo.buildconfig.demos"
+    version = "0.1.0-demo"
+}
