+++ demo-project/groovy/src/main/resources/file1.json
