+++ demo-project/groovy/src/main/resources/file2.json
