+++ demo-project/kts-js/src/test/kotlin/com/github/gmazzo/buildconfig/demos/kts_js/BuildConfigTest.kt
@@ -1,0 +1,19 @@
+package com.github.gmazzo.buildconfig.demos.kts_js
+
+import kotlin.js.Date
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertTrue
+
+class BuildConfigTest {
+
+    @Test
+    fun testBuildConfigProperties() {
+        assertEquals("kts-js", BuildConfig.APP_NAME)
+        assertEquals("Z3JhZGxlLWphdmEtYnVpbGRjb25maWctcGx1Z2lu", BuildConfig.APP_SECRET)
+        assertTrue(Date().getTime() >= BuildConfig.BUILD_TIME)
+        assertTrue(BuildConfig.FEATURE_ENABLED)
+        assertEquals(listOf(1, 2, 3, 4), BuildConfig.MAGIC_NUMBERS.toList())
+    }
+
+}
