+++ demo-project/kts/src/main/kotlin/com/github/gmazzo/buildconfig/demos/kts/SomeData.kt
@@ -1,0 +1,6 @@
+package com.github.gmazzo.buildconfig.demos.kts
+
+data class SomeData(
+    val value1: String,
+    val value2: Int
+)
