+++ demo-project/kts/src/main/resources/file1.json
