+++ demo-project/kts/src/main/resources/file2.json
