--- example-kts-multiplatform/src/commonTest/kotlin/com/github/gmazzo/example_kts_multiplatform/CommonBuildConfigTest.kt
@@ -1,13 +1,0 @@
-package com.github.gmazzo.example_kts_multiplatform
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class CommonBuildConfigTest {
-
-    @Test
-    fun testBuildConfigProperties() {
-        assertEquals("aCommonValue", BuildConfig.COMMON_VALUE)
-    }
-
-}
