--- example-kts-multiplatform/src/jsTest/kotlin/com/github/gmazzo/example_kts_multiplatform/BuildConfigTest.kt
@@ -1,15 +1,0 @@
-package com.github.gmazzo.example_kts_multiplatform
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class BuildConfigTest {
-
-    @Test
-    fun testBuildConfigProperties() {
-        assertEquals("js", JsMainBuildConfig.PLATFORM)
-        assertEquals("aCommonValue", BuildConfig.COMMON_VALUE)
-        assertEquals("aJsValue", JsMainBuildConfig.JS_VALUE)
-    }
-
-}
