--- example-kts/src/integrationTest/kotlin/BuildConfigITest.kt
@@ -1,12 +1,0 @@
-import com.github.gmazzo.example_kts.IntegrationTestBuildConfig
-import org.junit.Assert.assertEquals
-import org.junit.Test
-
-class BuildConfigITest {
-
-    @Test
-    fun testBuildConfigIntegrationTestProperties() {
-        assertEquals("aIntTestValue", IntegrationTestBuildConfig.INTEGRATION_TEST_CONSTANT)
-    }
-
-}
