--- example-kts/src/main/kotlin/com/github/gmazzo/example_kts/SomeData.kt
@@ -1,6 +1,0 @@
-package com.github.gmazzo.example_kts
-
-data class SomeData(
-    val value1: String,
-    val value2: Int
-)
