+++ gradle/libs.versions.toml
@@ -1,7 +1,10 @@
+[versions]
+kotlin = "1.8.20"
+
 [libraries]
 apache-commons-lang = { module = "org.apache.commons:commons-lang3", version = "3.12.0" }
 javapoet = { module = "com.squareup:javapoet", version = "1.13.0" }
-junit = { module = "junit:junit", version = "4.13.2" }
+kotlin-test = { module = "org.jetbrains.kotlin:kotlin-test", version.ref = "kotlin" }
 kotlinpoet = { module = "com.squareup:kotlinpoet", version = "1.10.2" }
 mockk = { module = "io.mockk:mockk", version = "1.13.5" }
 
@@ -9,4 +12,4 @@ mockk = { module = "io.mockk:mockk", version = "1.13.5" }
 gitVersioning = { id = "me.qoomon.git-versioning", version = "6.4.2" }
 gradle-pluginPublish = { id = "com.gradle.plugin-publish", version = "1.2.0" }
 jacoco-testkit = { id = "pl.droidsonroids.jacoco.testkit", version = "1.0.9" }
-kotlin = { id = "org.jetbrains.kotlin.jvm", version = "1.8.20" }
+kotlin = { id = "org.jetbrains.kotlin.jvm", version.ref = "kotlin" }
