+++ gradle/shared.settings.gradle.kts
@@ -1,5 +1,4 @@
 dependencyResolutionManagement {
-    repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
     repositories {
         mavenCentral()
     }
